# ui/src/api_document.json
{
    "openapi": "3.0.1",
    "info": {
      "title": "Xcianify.Presentation",
      "version": "1.0"
    },
    "paths": {
...
      "/api/items/{id}": {
        "get": {
          "tags": [
            "ItemMaster"
          ],
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "style": "simple",
              "schema": {
                "type": "integer",
                "format": "int32"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Success"
            }
          }
        },
        "put": {
          "tags": [
            "ItemMaster"
          ],
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "style": "simple",
              "schema": {
                "type": "integer",
                "format": "int32"
              }
            }
          ],
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateItemMasterDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateItemMasterDto"
                }
              },
              "application/*+json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateItemMasterDto"
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Success"
            }
          }
        },
        "delete": {
          "tags": [
            "ItemMaster"
          ],
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "style": "simple",
              "schema": {
                "type": "integer",
                "format": "int32"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Success"
            }
          }
        }
      },
      "/api/items/code/{itemCode}": {
        "get": {
          "tags": [
            "ItemMaster"
          ],
          "parameters": [
            {
              "name": "itemCode",
              "in": "path",
              "required": true,
              "style": "simple",
              "schema": {
                "type": "string"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Success"
            }
          }
        }
      },
...
      "/api/items/{itemMasterId}/sales-details": {
        "get": {
          "tags": [
            "ItemSalesDetail"
          ],
          "parameters": [
            {
              "name": "itemMasterId",
              "in": "path",
              "required": true,
              "style": "simple",
              "schema": {
                "type": "integer",
                "format": "int32"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Success"
            }
          }
        },
        "post": {
          "tags": [
            "ItemSalesDetail"
          ],
          "parameters": [
            {
              "name": "itemMasterId",
              "in": "path",
              "required": true,
              "style": "simple",
              "schema": {
                "type": "integer",
                "format": "int32"
              }
            }
          ],
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateItemSalesDetailDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateItemSalesDetailDto"
                }
              },
              "application/*+json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateItemSalesDetailDto"
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Success"
            }
          }
        }
      },
      "/api/items/{itemMasterId}/sales-details/{id}": {
        "put": {
          "tags": [
            "ItemSalesDetail"
          ],
          "parameters": [
            {
              "name": "itemMasterId",
              "in": "path",
              "required": true,
              "style": "simple",
              "schema": {
                "type": "integer",
                "format": "int32"
              }
            },
            {
              "name": "id",
              "in": "path",
              "required": true,
              "style": "simple",
              "schema": {
                "type": "integer",
                "format": "int32"
              }
            }
          ],
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateItemSalesDetailDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateItemSalesDetailDto"
                }
              },
              "application/*+json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateItemSalesDetailDto"
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Success"
            }
          }
        },
        "delete": {
          "tags": [
            "ItemSalesDetail"
          ],
          "parameters": [
            {
              "name": "itemMasterId",
              "in": "path",
              "required": true,
              "style": "simple",
              "schema": {
                "type": "integer",
                "format": "int32"
              }
            },
            {
              "name": "id",
              "in": "path",
              "required": true,
              "style": "simple",
              "schema": {
                "type": "integer",
                "format": "int32"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Success"
            }
          }
        }
      },
      "/api/items/{itemMasterId}/stock-analysis": {
        "get": {
          "tags": [
            "ItemStockAnalysis"
          ],
          "parameters": [
            {
              "name": "itemMasterId",
              "in": "path",
              "required": true,
              "style": "simple",
              "schema": {
                "type": "integer",
                "format": "int32"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Success"
            }
          }
        },
        "post": {
          "tags": [
            "ItemStockAnalysis"
          ],
          "parameters": [
            {
              "name": "itemMasterId",
              "in": "path",
              "required": true,
              "style": "simple",
              "schema": {
                "type": "integer",
                "format": "int32"
              }
            }
          ],
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateItemStockAnalysisDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateItemStockAnalysisDto"
                }
              },
              "application/*+json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateItemStockAnalysisDto"
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Success"
            }
          }
        }
      },
      "/api/items/{itemMasterId}/stock-analysis/{id}": {
        "put": {
          "tags": [
            "ItemStockAnalysis"
          ],
          "parameters": [
            {
              "name": "itemMasterId",
              "in": "path",
              "required": true,
              "style": "simple",
              "schema": {
                "type": "integer",
                "format": "int32"
              }
            },
            {
              "name": "id",
              "in": "path",
              "required": true,
              "style": "simple",
              "schema": {
                "type": "integer",
                "format": "int32"
              }
            }
          ],
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateItemStockAnalysisDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateItemStockAnalysisDto"
                }
              },
              "application/*+json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateItemStockAnalysisDto"
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Success"
            }
          }
        },
        "delete": {
          "tags": [
            "ItemStockAnalysis"
          ],
          "parameters": [
            {
              "name": "itemMasterId",
              "in": "path",
              "required": true,
              "style": "simple",
              "schema": {
                "type": "integer",
                "format": "int32"
              }
            },
            {
              "name": "id",
              "in": "path",
              "required": true,
              "style": "simple",
              "schema": {
                "type": "integer",
                "format": "int32"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Success"
            }
          }
        }
      },
...
    },
    "components": {
      "schemas": {
...
        "CreateItemSalesDetailDto": {
          "type": "object",
          "properties": {
            "itemMasterId": {
              "type": "integer",
              "format": "int32"
            },
            "sellingPrice": {
              "type": "number",
              "format": "double",
              "nullable": true
            },
            "currencyId": {
              "type": "integer",
              "format": "int32",
              "nullable": true
            },
            "isTaxInclusive": {
              "type": "boolean"
            },
            "discountPercentage": {
              "type": "number",
              "format": "double",
              "nullable": true
            },
            "minimumOrderQuantity": {
              "type": "number",
              "format": "double",
              "nullable": true
            },
            "leadTimeDays": {
              "type": "integer",
              "format": "int32",
              "nullable": true
            },
            "isActive": {
              "type": "boolean"
            },
            "notes": {
              "type": "string",
              "nullable": true
            }
          },
          "additionalProperties": false
        },
...
        "CreateItemStockAnalysisDto": {
          "type": "object",
          "properties": {
            "itemMasterId": {
              "type": "integer",
              "format": "int32"
            },
            "minimumStockLevel": {
              "type": "number",
              "format": "double",
              "nullable": true
            },
            "maximumStockLevel": {
              "type": "number",
              "format": "double",
              "nullable": true
            },
            "reorderLevel": {
              "type": "number",
              "format": "double",
              "nullable": true
            },
            "economicOrderQuantity": {
              "type": "number",
              "format": "double",
              "nullable": true
            },
            "leadTimeDays": {
              "type": "integer",
              "format": "int32",
              "nullable": true
            },
            "averageUsagePerDay": {
              "type": "number",
              "format": "double",
              "nullable": true
            },
            "lastStockCheckDate": {
              "type": "string",
              "format": "date-time",
              "nullable": true
            },
            "lastStockQuantity": {
              "type": "number",
              "format": "double",
              "nullable": true
            },
            "nextStockCheckDate": {
              "type": "string",
              "format": "date-time",
              "nullable": true
            },
            "isActive": {
              "type": "boolean"
            },
            "notes": {
              "type": "string",
              "nullable": true
            }
          },
          "additionalProperties": false
        },
...
        "UpdateItemSalesDetailDto": {
          "type": "object",
          "properties": {
            "sellingPrice": {
              "type": "number",
              "format": "double",
              "nullable": true
            },
            "currencyId": {
              "type": "integer",
              "format": "int32",
              "nullable": true
            },
            "isTaxInclusive": {
              "type": "boolean",
              "nullable": true
            },
            "discountPercentage": {
              "type": "number",
              "format": "double",
              "nullable": true
            },
            "minimumOrderQuantity": {
              "type": "number",
              "format": "double",
              "nullable": true
            },
            "leadTimeDays": {
              "type": "integer",
              "format": "int32",
              "nullable": true
            },
            "isActive": {
              "type": "boolean",
              "nullable": true
            },
            "notes": {
              "type": "string",
              "nullable": true
            }
          },
          "additionalProperties": false
        },
...
        "UpdateItemStockAnalysisDto": {
          "type": "object",
          "properties": {
            "minimumStockLevel": {
              "type": "number",
              "format": "double",
              "nullable": true
            },
            "maximumStockLevel": {
              "type": "number",
              "format": "double",
              "nullable": true
            },
            "reorderLevel": {
              "type": "number",
              "format": "double",
              "nullable": true
            },
            "economicOrderQuantity": {
              "type": "number",
              "format": "double",
              "nullable": true
            },
            "leadTimeDays": {
              "type": "integer",
              "format": "int32",
              "nullable": true
            },
            "averageUsagePerDay": {
              "type": "number",
              "format": "double",
              "nullable": true
            },
            "lastStockCheckDate": {
              "type": "string",
              "format": "date-time",
              "nullable": true
            },
            "lastStockQuantity": {
              "type": "number",
              "format": "double",
              "nullable": true
            },
            "nextStockCheckDate": {
              "type": "string",
              "format": "date-time",
              "nullable": true
            },
            "isActive": {
              "type": "boolean",
              "nullable": true
            },
            "notes": {
              "type": "string",
              "nullable": true
            }
          },
          "additionalProperties": false
        },
...
      },
      "securitySchemes": {
        "Bearer": {
          "type": "apiKey",
          "description": "JWT Authorization header using the Bearer scheme. Enter 'Bearer' [space] and then your token in the text input below.",
          "name": "Authorization",
          "in": "header"
        }
      }
    },
    "security": [
      {
        "Bearer": [ ]
      }
    ]
  }